+++ app/Http/Controllers/UserController.php
@@ -7,14 +7,17 @@ use App\Http\Requests\UpdateUserRequest;
 use App\Http\Resources\UserResource;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class UserController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAny', User::class);
         $users = User::query()
             ->when($request->has('email'), function($query) use ($request) {
                 $query->where('email', 'like', '%' . $request->input('email') . '%');
@@ -32,6 +35,7 @@ class UserController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', User::class);
         return $this->index(new Request());
     }
 
@@ -40,6 +44,7 @@ class UserController extends Controller
      */
     public function store(StoreUserRequest $request)
     {
+        $this->authorize('create', User::class);
         try {
             User::create($request->validated());
             return redirect()->route('users.index')->with('flash.success', 'User created successfully.');
@@ -61,6 +66,7 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        $this->authorize('update', $user);
         return inertia('Users/Index', [
             'user' => new UserResource($user),
         ]);
@@ -71,6 +77,7 @@ class UserController extends Controller
      */
     public function update(UpdateUserRequest $request, User $user)
     {
+        $this->authorize('update', $user);
         try {
             $user->update($request->validated());
             return redirect()->route('users.index')->with('flash.success', 'User updated successfully.');
@@ -84,11 +91,13 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
-       try {
+        $this->authorize('delete', $user);
+        
+        try {
             $user->delete();
             return redirect()->route('users.index')->with('flash.success', 'User deleted successfully.');
         } catch (\Exception $e) {
             return redirect()->back()->with('flash.error', 'Error deleting user.');
-       }
+        }
     }
 }
